Multi-selection: allow shift+click from single block selection
